Cleanup SHA-3 Clang workaround
